[chore]: refactor map renderer
